requirements.txt
g4f
matplotlib
python-telegram-bot
Requests
SQLAlchemy
